refactor(workflow): Move code to `replay_advance`
